+++ database/migrations/2017_06_06_134041_add_label_to_entity_template_fields.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddLabelToEntityTemplateFields extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('entity_template_fields', function (Blueprint $table) {
+            $table->string('label')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('entity_template_fields', function (Blueprint $table) {
+            $table->dropColumn('label');
+        });
+    }
+}